requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imutils==0.5.4
intel-openmp==2021.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
orjson==3.10.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.2.0
psutil==5.9.8
psycopg2==2.9.10
py-cpuinfo==9.0.0
pydantic==2.7.4
pydantic-settings==2.6.0
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.37.2
sympy==1.12
tbb==2021.11.0
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
ultralytics-thop==0.2.8
urllib3==2.2.1
uvicorn==0.32.0
watchfiles==0.22.0
websockets==12.0
